divide freestyle.util into different packages
