+++ src/main/scala/freestyle/schema.scala
@@ -17,6 +17,7 @@
 package skeuomorph
 package freestyle
 
+import qq.droste.Algebra
 import cats.Functor
 import cats.data.NonEmptyList
 
@@ -60,6 +61,37 @@ object Schema {
       case TSum(name, fields)     => TSum(name, fields)
       case TProduct(name, fields) => TProduct(name, fields.map(field => field.copy(tpe = f(field.tpe))))
     }
+  }
+
+  def render: Algebra[Schema, String] = Algebra {
+    case TNull()          => "Null"
+    case TDouble()        => "Double"
+    case TFloat()         => "Float"
+    case TInt()           => "Int"
+    case TLong()          => "Long"
+    case TBoolean()       => "Boolean"
+    case TString()        => "String"
+    case TByteArray()     => "Array[Byte]"
+    case TNamedType(name) => name
+    case TOption(value)   => s"Option[$value]"
+    case TMap(value)      => s"Map[String, $value]"
+    case TList(value)     => s"List[$value]"
+    case TRequired(value) => value
+    case TCoproduct(invariants) =>
+      invariants.toList.mkString("Cop[", " :: ", ":: TNil]")
+    case TSum(name, fields) =>
+      val printFields = fields.map(f => s"case object $f extends $name").mkString("\n  ")
+      s"""
+sealed trait $name
+object $name {
+  $printFields
+}
+"""
+    case TProduct(name, fields) =>
+      val printFields = fields.map(f => s"${f.name}: ${f.tpe}").mkString(", ")
+      s"""
+@message case class $name($printFields)
+"""
   }
 
 }
